+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file. The format 
 ### Changed
 - Switch to a `src/`-layout.
 - Import base-element class as `element.Element()`.
+- Auto-detect min/max-coordinates of mesh-points for loadcases if `dof.uniaxial(right=None)`.
 
 ### Fixed
 - Fix rotation matrix for a rotation around the y-axis.